Alterando tipo do atributo Peso (int para numeric) e alteração dos pesos dos itens iniciais

+++ mod_5/prev/trigger_functions/atualizar_volume_ocupado_mochila.sql
@@ -5,7 +5,7 @@ RETURNS TRIGGER AS $atualizar_volume_ocupado_mochila$
 DECLARE 
 
   volume_ocupado_atual INTEGER;
-  peso_item INTEGER;
+  peso_item NUMERIC(4,2);
   tableName VARCHAR(50);
   capacidade_mochila INTEGER;
